MUZIMA-607: Enhace resolution of potential duplicate registrations
The user has the ability to either merge the incoming record or creating a new record.
In case the user chooses to merge the selected details will be updated on the existing patient.
Otherwise if the user chooses to create a new record, the system ensures the IDs if present are
not the same prompting the user to change if that is the case. Moreover choosing to create a new
patient record effectively means the system is no longer going to check for potential matches. This
is achieved by setting a flag on the payload.